Move UDE handling to a its own file, adjust wording
